Remove article and page source `<link>`'s
This feature was causing the plain text URLs of my articles to appear
in search results (at least in Duck Duck Go, and I assume in others as
well).

It's not possible to add a `<meta name="robots" content="noindex">` tag
to these pages because they're plain text, not HTML.

Nor can I add the `X-Robots-Tag` HTTP header: my web host (GitHub Pages)
doesn't give me control of headers.

I thought this feature was a neat easter egg, but it's actually just
pointless trouble. Just remove it.

I'm leaving the ability for users to add source links to their article
and page taglines or footers, for backwards-compatibility if nothing
else, but adding a warning to the docs and removing it from the default
article footer.