Remove version number from README
Edit the README to avoid actually including the latest version number.
This removes the need to update the version number in the README when
releasing a new version.

This also allows me to remove the `make readme` command and a
`readme.yml` workflow that was sending automated PRs to retrospectively
update the version number in the README after doing a release. These
updated PRs would have been broken since this repo was moved into the
pelican-themes organization anyway: it used a GitHub App that only had
access to my GitHub organization.

This also means that the README can now be included in the release
archives (since it won't contain an outdated version number).